requirements.txt
# Data Archaeologist Framework Dependencies
# Professional database analysis and reverse engineering tools

# Core Database Connectivity
psycopg2-binary>=2.9.10
sqlalchemy>=2.0.20

# Data Analysis and Processing
pandas>=2.0.0
numpy>=1.24.0

# Output and Reporting
openpyxl>=3.1.0

# Configuration Management
pyyaml>=6.0

# Modern Task Runner and CLI
click>=8.1.0
rich>=13.0.0

# Development and Quality Tools (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
